+++ frontend/src/components/ProductQualityVisualization.vue
@@ -18,7 +18,7 @@
           <ul v-for="lvl in allLevels" :key="lvl">
             <li class="header">Level {{ lvl }}</li>
             <li
-              v-for="product in filterProductsByLvl(lvl)"
+              v-for="product in filterProductsByLvl(props, lvl)"
               :key="product.id"
               class="qualityitem"
               :class="{
@@ -67,13 +67,10 @@ import {
   level2ContainsErrors,
   level2containsWarningsOrInfo,
   onlyLegacyLevel2,
+  filterProductsByLvl,
+  type Props,
 } from "@/lib/ProductAvailabilityTools";
 import DateVisualization from "./DateVisualization.vue";
-import type { DataStatus } from "@/lib/DataStatusParser";
-
-export interface Props {
-  dataStatus: DataStatus;
-}
 
 const props = defineProps<Props>();
 const dates = computed(() =>
@@ -85,13 +82,6 @@ const dates = computed(() =>
 );
 
 const allLevels = computed(() => Array.from(new Set(Object.values(props.dataStatus.lvlTranslate))).sort());
-
-function filterProductsByLvl(lvl: string) {
-  if (!props.dataStatus.allProducts) return null;
-  return props.dataStatus.allProducts
-    .filter(({ id }) => props.dataStatus.lvlTranslate[id] == lvl && id != "model")
-    .filter(({ experimental }) => !experimental);
-}
 
 function createColorClass(products: ProductLevels): ColorClass {
   if (noData(products)) return "no-data";
